Publish version 3
The main change is to employ a coroutine to scan the grid. This is necessary to scan larger grids, as there are limits to how many instructions a single execution can perform.

The same problem exists for rendering, which is not yet addressed in this version.

The current version was tested with a 5500 block ship.